Implement semantics flag
The default semantics changed to Shen & Eiter epistemic semantics under
FLP answer set semantics where nested expressions of form `not not a`
evaluate to `a`.

With SE epistemic semantics under FLP semantics, the following program has
the single world view {{}} wrt the guess G = {not K a}:

a ← M a.
← K a.

The epistemic reduct wrt G under FLP semantics of this program is:

a ← a.

The guess G is not consistent wit the answer sets of the previously used
epistemic reduct under Kahl semantics:

a ← not not a.

This reduct has two answersets {}, {a}, because Kahl defines nested
expressions as in [1], but M a is guessed to be false.

[1] Lifschitz, V., Tang, L. R., & Turner, H. (1999). Nested Expressions in Logic Programs